# backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "娱乐中心系统API",
    "description": "\n        ## 娱乐中心系统API文档\n        \n        这是一个基于FastAPI的娱乐中心系统，提供以下功能：\n        \n        ### 🎮 游戏模块\n        - **刮刮乐游戏**: 多种主题的刮刮乐游戏\n        - **老虎机游戏**: 经典和现代老虎机\n        - **幸运大转盘**: 多种类型的转盘游戏\n        \n        ### 👤 用户管理\n        - 用户注册和登录\n        - 积分管理\n        - 游戏历史记录\n        \n        ### 📊 数据统计\n        - 用户游戏统计\n        - 排行榜系统\n        - 实时游戏状态\n        \n        ### 🔧 管理后台\n        - 用户管理\n        - 游戏配置\n        - 数据分析\n        \n        ### 🔐 认证方式\n        使用JWT Bearer Token进行认证。在请求头中添加：\n        ```\n        Authorization: Bearer <your_token>\n        ```\n        \n        ### 📱 响应格式\n        所有API响应都采用JSON格式，错误响应包含以下字段：\n        - `error`: 是否为错误\n        - `message`: 错误消息\n        - `status_code`: HTTP状态码\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/stats/user/stats": {
      "get": {
        "tags": [
          "数据统计"
        ],
        "summary": "Get User Stats",
        "description": "获取用户游戏统计",
        "operationId": "get_user_stats_api_stats_user_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGameStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stats/leaderboard/credits": {
      "get": {
        "tags": [
          "数据统计"
        ],
        "summary": "Get Credits Leaderboard",
        "description": "获取积分排行榜",
        "operationId": "get_credits_leaderboard_api_stats_leaderboard_credits_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/leaderboard/total-win": {
      "get": {
        "tags": [
          "数据统计"
        ],
        "summary": "Get Total Win Leaderboard",
        "description": "获取总赢取排行榜",
        "operationId": "get_total_win_leaderboard_api_stats_leaderboard_total_win_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/leaderboard/win-rate": {
      "get": {
        "tags": [
          "数据统计"
        ],
        "summary": "Get Win Rate Leaderboard",
        "description": "获取胜率排行榜（需要最少游戏次数）",
        "operationId": "get_win_rate_leaderboard_api_stats_leaderboard_win_rate_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "min_games",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10,
              "title": "Min Games"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/analysis": {
      "get": {
        "tags": [
          "数据统计"
        ],
        "summary": "Get Game Analysis",
        "description": "获取游戏分析数据（需要管理员权限或自己的数据）",
        "operationId": "get_game_analysis_api_stats_analysis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Game Type"
            }
          },
          {
            "name": "template_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameAnalysisResponse": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "total_players": {
            "type": "integer",
            "title": "Total Players"
          },
          "total_games": {
            "type": "integer",
            "title": "Total Games"
          },
          "total_revenue": {
            "type": "integer",
            "title": "Total Revenue"
          },
          "total_payout": {
            "type": "integer",
            "title": "Total Payout"
          },
          "profit_margin": {
            "type": "number",
            "title": "Profit Margin"
          },
          "popular_games": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Popular Games"
          },
          "hourly_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Hourly Distribution"
          },
          "daily_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Daily Distribution"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total_players",
          "total_games",
          "total_revenue",
          "total_payout",
          "profit_margin",
          "popular_games",
          "hourly_distribution",
          "daily_distribution"
        ],
        "title": "GameAnalysisResponse",
        "description": "游戏分析响应"
      },
...
      "GameStatsResponse": {
        "properties": {
          "total_games": {
            "type": "integer",
            "title": "Total Games"
          },
          "total_bet": {
            "type": "integer",
            "title": "Total Bet"
          },
          "total_win": {
            "type": "integer",
            "title": "Total Win"
          },
          "net_result": {
            "type": "integer",
            "title": "Net Result"
          },
          "win_rate": {
            "type": "number",
            "title": "Win Rate"
          },
          "favorite_game": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorite Game"
          }
        },
        "type": "object",
        "required": [
          "total_games",
          "total_bet",
          "total_win",
          "net_result",
          "win_rate",
          "favorite_game"
        ],
        "title": "GameStatsResponse",
        "description": "游戏统计响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LeaderboardEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "game_count": {
            "type": "integer",
            "title": "Game Count"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "user_id",
          "username",
          "value",
          "game_count"
        ],
        "title": "LeaderboardEntry",
        "description": "排行榜条目"
      },
      "LeaderboardResponse": {
        "properties": {
          "leaderboard_type": {
            "type": "string",
            "title": "Leaderboard Type"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "user_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Rank"
          },
          "user_value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Value"
          }
        },
        "type": "object",
        "required": [
          "leaderboard_type",
          "entries"
        ],
        "title": "LeaderboardResponse",
        "description": "排行榜响应"
      },
...
      "UserGameStatsResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "overall_stats": {
            "$ref": "#/components/schemas/GameStatsResponse"
          },
          "game_type_stats": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GameStatsResponse"
            },
            "type": "object",
            "title": "Game Type Stats"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "overall_stats",
          "game_type_stats"
        ],
        "title": "UserGameStatsResponse",
        "description": "用户游戏统计响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "开发服务器"
    },
    {
      "url": "https://api.entertainment-center.com",
      "description": "生产服务器"
    }
  ],
...
}